formulario de envio actualizado 13/08

+++ GraficaSolemno_gitIgnore/CapaDatos/CDDetallesEnvio.vb
@@ -5,7 +5,7 @@ Public Class CDDetallesEnvio
     Dim oCDConexion As New CDConexion
 
     Public Function InformacionDeEnvio(ByVal id As Integer) As DataTable
-        consulta = "select * from DetallesEnvio where detallesenvio.IDPedido=" & id
+        consulta = "select DetallesEnvio.IDEnvio,DetallesEnvio.IDPedido, DetallesEnvio.IDCliente ,provincias.nombre as 'Provincia', localidad.nombre as 'Localidad', DetallesEnvio.Domicilio, DetallesEnvio.NroCalle, DetallesEnvio.Dpto, DetallesEnvio.CP, DetallesEnvio.Barrio, DetallesEnvio.Transporte, DetallesEnvio.NroSeguimiento, DetallesEnvio.PrecioEnvio, DetallesEnvio.FechaDespacho from DetallesEnvio,provincias,localidad where detallesenvio.IDProvincia = provincias.IDProvincia And DetallesEnvio.IDLocalidad = Localidad.IDLocalidad And detallesenvio.IDPedido = " & id
         Return oCDConexion.MostrarTablaModificada(consulta)
     End Function
    
@@ -21,7 +21,7 @@ Public Class CDDetallesEnvio
             .Add("@IDProvincia", SqlDbType.Int).Value = datos.Provincia
             .Add("@IDLocalidad", SqlDbType.Int).Value = datos.Localidad
             .Add("@Domicilio", SqlDbType.VarChar).Value = datos.Domicilio
-            .Add("@NroCalle", SqlDbType.Int).Value = SetNullValues(datos.NroCalle)
+            .Add("@NroCalle", SqlDbType.VarChar).Value = SetNullValues(datos.NroCalle)
             .Add("@Dpto", SqlDbType.VarChar).Value = SetNullValues(datos.Dpto)
             .Add("@CP", SqlDbType.VarChar).Value = SetNullValues(datos.CP)
             .Add("@Barrio", SqlDbType.VarChar).Value = datos.Barrio
@@ -59,13 +59,18 @@ Public Class CDDetallesEnvio
     End Sub
     Public Sub EliminarInformacionEnvio(ByVal id As Integer)
         oCDConexion.Conectar()
-        consulta = "delete from DetallesEnvio where IDPedido= @IDPedido"
+        Dim dt As New DataTable
+        dt = InformacionDeEnvio(id)
+        If dt.Rows.Count > 0 Then
+            consulta = "delete from DetallesEnvio where IDPedido= @IDPedido"
 
-        Dim cmd As New SQLiteCommand(consulta, oCDConexion.con)
-        With cmd.Parameters
-            .Add("@IDPedido", SqlDbType.Int).Value = id
-        End With
-        cmd.ExecuteNonQuery()
+            Dim cmd As New SQLiteCommand(consulta, oCDConexion.con)
+            With cmd.Parameters
+                .Add("@IDPedido", SqlDbType.Int).Value = id
+            End With
+            cmd.ExecuteNonQuery()
+        End If
+     
         oCDConexion.Desconectar()
     End Sub
     Public Function SetNullValues(ByVal value As Object) As Object
